+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 ## [UNRELEASED] - 2024-11-03
 
 ### Added
-- Debug support for most necessary functions with a wrapper
+- Debug support for most necessary functions with a wrapper.
+- Card for showing the node weight currently in network.
 
 ### Changed
 - Better logging when fetching `active_nets()`